==> 3_SHARE_TRACE/07_SHARE_TRACE.rescore_programs.ipynb <==
# Chromvar rescore topics

%load_ext autoreload
%autoreload 2
    
import os
os.environ["RAY_TQDM_PATCH_PRINT"]="0"
from umap import *
import numpy as np
import scprinter as scp
import pandas as pd
import seaborn as sns
import matplotlib.pyplot as plt
import time
import pandas as pd
import numpy as np
import os
import pickle
import torch
import matplotlib as mpl
import scanpy as sc
import anndata as ad
from scmallet import *
mpl.rcParams['pdf.fonttype'] = 42

import scanpy as sc
import anndata
import cupy as cp
import cupyx as cpx
import time

import warnings
warnings.filterwarnings("ignore")
import rmm
from rmm.allocators.cupy import rmm_cupy_allocator
rmm.reinitialize(
    managed_memory=True, # Allows oversubscription
    pool_allocator=True, # default is False
    devices=0, # GPU device IDs to register. By default registers only GPU 0.
)

cp.cuda.set_allocator(rmm_cupy_allocator)

work_dir = '/data/rzhang/data_15/colitisDSS/tissue'
meta = 'DSS-prim-merge10-ATAC.qc_stats.singlet_epi_only.txt'
frags = 'DSS-prim-merge10-ATAC.reshift.fragments.tsv.gz'
peak = 'DSS-prim-merge10-ATAC.cleanPeaks.fixedwidthpeaks_800bp_300reSized.bed'

atac_adata = ad.read_h5ad(f'{work_dir}/cellxpeak.h5ad')

rna_adata = sc.read_mtx('/data/rzhang/data_15/colitisDSS/tissue/joint_topic/DSS-prim-merge10-RNA.raw_UMI.merged.converted_barcodes.mtx.gz')
rna_adata = ad.AnnData(X = rna_adata.X.T)
bc = [line.strip() for line in open('/data/rzhang/data_15/colitisDSS/tissue/joint_topic/DSS-prim-merge10-RNA.raw_UMI.merged.converted_barcodes.barcodes.list', 'r').readlines()]
rna_adata.obs.index = bc
bc = [line.strip() for line in open('/data/rzhang/data_15/colitisDSS/tissue/joint_topic/DSS-prim-merge10-RNA.raw_UMI.merged.converted_barcodes.genes.list', 'r').readlines()]
rna_adata.var.index = bc

meta_cell = pd.read_csv('/data/rzhang/data_15/colitisDSS/tissue/joint_topic/DSS-prim-merge10-ATAC.qc_stats.singlet.txt', sep='\t')shared = list(set(meta_cell['barcode']) &  set(atac_adata.obs.index)  & set(rna_adata.obs.index))
atac_adata = atac_adata[shared].copy()
rna_adata = rna_adata[shared].copy()

atac_feats = pd.read_csv("/data/rzhang/data_15/colitisDSS/tissue/joint_chromvar/DSS-prim-merge10.ATAC.top_features.1000.annot.txt.gz", sep='\t')
atac_bg = pd.read_csv("/data/rzhang/data_15/colitisDSS/tissue/joint_chromvar/DSS-prim-merge10.atac_bg.csv.gz", header=None)

atac_feat_weighted = pd.read_csv("/data/rzhang/data_15/colitisDSS/tissue/joint_topic/topic90_region_topics_ATAC.txt.gz", sep='\t', index_col=0)

atac_adata.varm['bg_peaks'] = atac_bg.values.astype('int')  - 1  # minus 1 because R starts at 1, python starts at 0
atac_adata.varm['motif_match'] = atac_feats.values
atac_adata.uns['motif_name'] = atac_feats.columns

chromvar_atac_surya = scp.chromvar.compute_deviations(atac_adata, chunk_size=50000, device='cuda')

atac_adata.varm['motif_match'] = atac_feat_weighted.values
chromvar_atac_raw_weighted = scp.chromvar.compute_deviations(atac_adata, chunk_size=50000, device='cuda')

# Now for RNA

rna_feats = pd.read_csv("/data/rzhang/data_15/colitisDSS/tissue/joint_chromvar/DSS-prim-merge10.RNA.top_features.100.annot.txt.gz", sep='\t')
rna_bg = pd.read_csv("/data/rzhang/data_15/colitisDSS/tissue/joint_chromvar/DSS-prim-merge10.rna_bg.csv.gz", header=None)

rna_feat_weighted = pd.read_csv("/data/rzhang/data_15/colitisDSS/tissue/joint_topic/topic90_region_topics_RNA.txt.gz", sep='\t', index_col=0)

rna_adata = rna_adata[:, rna_feat_weighted.index].copy()

rna_adata.varm['bg_peaks'] = rna_bg.values.astype('int')  - 1
rna_adata.varm['motif_match'] = rna_feat_weighted.loc[rna_adata.var.index].values
rna_adata.uns['motif_name'] = rna_feat_weighted.columns

rna_adata.X = rna_adata.X.tocsr()

chromvar_rna = scp.chromvar.compute_deviations(rna_adata, chunk_size=50000, device='cuda')

chromvar_atac_df = pd.DataFrame(chromvar_atac_raw_weighted.X, index=chromvar_atac.obs.index, columns=chromvar_atac.var.index)
chromvar_rna_df = pd.DataFrame(chromvar_rna.X, index=chromvar_rna.obs.index, columns=chromvar_rna.var.index)

chromvar_atac_df

chromvar_rna_df

chromvar_atac_df.to_csv('/data/rzhang/data_15/colitisDSS/tissue/joint_chromvar/chromvar_atac.sep.weighted.txt.gz', sep='\t')
chromvar_rna_df.to_csv('/data/rzhang/data_15/colitisDSS/tissue/joint_chromvar/chromvar_rna.sep.weighted.txt.gz', sep='\t')
